% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bppinmf}
\alias{bppinmf}
\title{Block Principal Pivoted Iterative Non-Negative Matrix Factorization}
\usage{
bppinmf(
  objectList,
  k,
  lambda,
  maxIter,
  thresh,
  verbose = TRUE,
  Hinit = NULL,
  Vinit = NULL,
  Winit = NULL
)
}
\arguments{
\item{Ei}{List of datasets in dense matrix form.}
}
\value{
The calculated solution matrix in dense form.
}
\description{
Use the BPP algorithm to iteratively factor the given datasets.
}
\examples{
\dontshow{if (require("Matrix")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bppinmf(rsparsematrix(nrow=20,ncol=20,nnz=10), Matrix(runif(n=200,min=0,max=2),20,10))
\dontshow{\}) # examplesIf}
}
