% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bppnmf}
\alias{bppnmf}
\title{Alternating  Nonnegative Least Squares with Block Principal Pivoting NMF}
\usage{
bppnmf(x, k, niter, W_init = NULL, H_init = NULL)
}
\arguments{
\item{x}{Input matrix for factorization}

\item{k}{Factor matrix rank}

\item{niter}{Maximum number of nmf iterations}

\item{W_init}{Initial left-hand factor matrix (Optional)}

\item{H_init}{Initial right-hand factor matrix (Optional)}
}
\value{
The calculated factor matrices as an Rcpp::List
}
\description{
Use the ANLS-BPP algorithm to factor a given matrix at the given rank.
}
\examples{
\dontshow{if (require("Matrix")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
munmf(rsparsematrix(nrow = 100, ncol = 100, nnz = 10, symmetric = TRUE), 10, 10)
\dontshow{\}) # examplesIf}
}
