% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bppnnls}
\alias{bppnnls}
\title{Block Principal Pivoted Non-Negative Least Squares}
\usage{
bppnnls(C, B)
}
\arguments{
\item{C}{Input factor dense matrix}

\item{B}{Input sparse matrix}
}
\value{
The calculated solution matrix in dense form.
}
\description{
Use the BPP algorithm to get the nonnegative least squares solution for the given matrices.
}
