% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctrl.sparse}
\alias{ctrl.sparse}
\alias{stim.sparse}
\title{Example single-cell transcriptomic data in sparse form}
\format{
An object of class \code{dgCMatrix} with 173 rows and 300 columns.

An object of class \code{dgCMatrix} with 173 rows and 300 columns.
}
\source{
https://www.nature.com/articles/nbt.4042
}
\usage{
ctrl.sparse

stim.sparse
}
\description{
The two datasets, namingly \code{ctrl.sparse} and \code{stim.sparse}, are
single-cell transcriptomic data preprocessed and subsampled from the study
of Hyun Min Kang and et al., Nat Biotech., 2018. The raw datasets were two
sparse matrices of integer values indicating the counts of genes (rows) per
cell (columns). We normalized each column of both matrices by its library
size (sum), and selected common variable genes across the datasets. Finally,
we scaled the genes without centering them, in order to keep the
non-negativity. The processed datasets were then randomly subsampled for a
minimal example.
}
\keyword{datasets}
