% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.H5SpMat}
\alias{dim.H5SpMat}
\alias{dim<-.H5SpMat}
\title{Retrieve the dimension of H5SpMat argument list}
\usage{
\method{dim}{H5SpMat}(x)

\method{dim}{H5SpMat}(x) <- value
}
\arguments{
\item{x}{H5SpMat argument list object}

\item{value}{Numeric vector of two, for number of rows and number of columns.}
}
\value{
Retriever returns a vector of two (nrow and ncol), setter sets the
value of that in the argument list.
}
\description{
Retrieve the dimension of H5SpMat argument list
}
\examples{
h <- H5SpMat(system.file("extdata/ctrl_sparse.h5", package = "RcppPlanc"),
             "data", "indices", "indptr", 173, 300)
dim(h)
nrow(h)
ncol(h)
dim(h) <- c(200, 200)
h
}
