% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format.H5Mat}
\alias{format.H5Mat}
\title{Prepare character information of a H5Mat object}
\usage{
\method{format}{H5Mat}(x, ...)
}
\arguments{
\item{x}{H5Mat argument list object}

\item{...}{Not used.}
}
\value{
A character scalar of the displayed message
}
\description{
Prepare character information of a H5Mat object
}
\examples{
h <- H5Mat(system.file("extdata/ctrl_dense.h5", package = "RcppPlanc"),
           "data")
format(h)
}
