% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format.H5SpMat}
\alias{format.H5SpMat}
\title{prepare character information of a H5SpMat object}
\usage{
\method{format}{H5SpMat}(x, ...)
}
\arguments{
\item{x}{H5SpMat argument list object}

\item{...}{Not used.}
}
\value{
A character scalar of the displayed message
}
\description{
prepare character information of a H5SpMat object
}
\examples{
h <- H5SpMat(system.file("extdata/ctrl_sparse.h5", package = "RcppPlanc"),
             "data", "indices", "indptr", 173, 300)
format(h)
}
