% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{H5Mat}
\alias{H5Mat}
\title{Create argument list for constructing HDF5 based dense matrix}
\usage{
H5Mat(filename, dataPath)
}
\arguments{
\item{filename}{Filename of the HDF5 file}

\item{dataPath}{Path in the HDF5 file that points to a 2D dense matrix}
}
\value{
H5Mat object, can be used like a list
}
\description{
Create argument list for constructing HDF5 based dense matrix
}
