% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{H5SpMat}
\alias{H5SpMat}
\title{Create argument list for constructing HDF5 based CSC sparse matrix}
\usage{
H5SpMat(filename, valuePath, rowindPath, colptrPath, nrow, ncol)
}
\arguments{
\item{filename}{Filename of the HDF5 file}

\item{valuePath}{Path in the HDF5 file that points to the 1D value vector of
the sparse matrix.}

\item{rowindPath}{Path in the HDF5 file that points to the 1D rowind vector
of the sparse matrix.}

\item{colptrPath}{Path in the HDF5 file that points to the 1D colptr vector
of the sparse matrix.}

\item{nrow}{Integer, number of rows of the sparse matrix.}

\item{ncol}{Integer, number of columns of the sparse matrix.}
}
\value{
H5SpMat object, can be used like a list
}
\description{
Create argument list for constructing HDF5 based CSC sparse matrix
}
