% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.H5Mat}
\alias{print.H5Mat}
\title{Show info of H5Mat object}
\usage{
\method{print}{H5Mat}(x, ...)
}
\arguments{
\item{x}{H5Mat object}

\item{...}{not used}
}
\description{
Show info of H5Mat object
}
