% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.H5SpMat}
\alias{print.H5SpMat}
\title{Show info of H5SpMat object}
\usage{
\method{print}{H5SpMat}(x, ...)
}
\arguments{
\item{x}{H5SpMat object}

\item{...}{not used}
}
\description{
Show info of H5SpMat object
}
