% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppplanc_bppnnls}
\alias{rcppplanc_bppnnls}
\title{Block Principal Pivoted Non-Negative Least Squares}
\usage{
rcppplanc_bppnnls(A, B)
}
\arguments{
\item{A}{Input sparse matrix}

\item{B}{Input factor dense matrix}
}
\value{
The calculated solution matrix in dense form.
}
\description{
Use the BPP algorithm to get the nonnegative least squares solution for the given matrices.
}
\examples{
\dontshow{if (require("Matrix")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rcppplanc_bppnnls(rsparsematrix(nrow=20,ncol=20,nnz=10), matrix(runif(n=200,min=0,max=2),20,10))
\dontshow{\}) # examplesIf}
}
