% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppplanc_gnsymnmf}
\alias{rcppplanc_gnsymnmf}
\title{Gauss-Newton using Conjugate Gradients NMF}
\usage{
rcppplanc_gnsymnmf(x, k, niter)
}
\arguments{
\item{x}{Input matrix for factorization}

\item{k}{Factor matrix rank}

\item{niter}{Maximum number of nmf iterations}
}
\value{
The calculated factor matrices as an Rcpp::list
}
\description{
Use the Gauss-Newton algorithm to factor a given matrix at the given rank.
}
\examples{
\dontshow{if (require("Matrix")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rcppplanc_gnsymnmf(rsparsematrix(nrow = 100, ncol = 100, nnz = 10, symmetric = TRUE), 10, 10)
\dontshow{\}) # examplesIf}
}
