% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uinmf}
\alias{uinmf}
\title{Block Principal Pivoted Itegrative Non-Negative Matrix Factorization with Unshared features}
\usage{
uinmf(objectList, unsharedList, k, lambda = 5, niter = 30L, verbose = TRUE)
}
\arguments{
\item{objectList}{List of matrices of datasets. Can be of class matrix or dgCMatrix.
All should have the same number of rows.}

\item{unsharedList}{List of matrices of unshared features. Should have the same
number of matrices as \code{objectList}. The number of columns should match the
number of columns in the corresponding matrix in \code{objectList}. Matrix with
zero number of rows is allowed. The class of the matrices should be the same as
the matrices in \code{objectList}.}

\item{k}{Number of factors to factorize the matrices into.}

\item{lambda}{Regularization parameter. Default \code{5}}

\item{niter}{Number of ANLS iterations to run. Default \code{30}}

\item{verbose}{Logical,. Whether to print the progress of the algorithm. Default \code{TRUE}}
}
\value{
A list of factorization result matrices, including the elements:
\code{H} list of H matrices, \code{V} list of V matrices, \code{W} W matrix,
\code{U} list of U matrices, \code{objErr} objective error.
}
\description{
Use the BPP algorithm to iteratively factor the given datasets.
}
