% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.H5Mat}
\alias{print.H5Mat}
\title{Show information of a H5Mat object}
\usage{
\method{print}{H5Mat}(x, ...)
}
\arguments{
\item{x}{H5Mat argument list object}

\item{...}{Not used.}
}
\value{
NULL. Information displayed.
}
\description{
Show information of a H5Mat object
}
\examples{
h <- H5Mat(system.file("extdata/ctrl_dense.h5", package = "RcppPlanc"),
           "data")
print(h)
}
