% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.H5SpMat}
\alias{print.H5SpMat}
\title{Show information of a H5SpMat object}
\usage{
\method{print}{H5SpMat}(x, ...)
}
\arguments{
\item{x}{H5SpMat argument list object}

\item{...}{Not used.}
}
\value{
NULL. Information displayed.
}
\description{
Show information of a H5SpMat object
}
\examples{
h <- H5SpMat(system.file("extdata/ctrl_sparse.h5", package = "RcppPlanc"),
             "data", "indices", "indptr", 173, 300)
print(h)
}
